reverted back aac2spdif to the simple filter and add aacspdifbin istead.
to co-operate with non-digital audio device.
previously, aacparse & aac2pdif were always selected by decodebin2
over faad etc., regardless of the availability of S/PDIF interface.
it's because the alsasink is always regarded to include "audio/x-iec958",
as the auto-plugging system just uses element factories and template caps,
and does not call gst_pad_getcaps() dynamically on an element instance.
In addition, the ranks of aacparse and aac2spdif was set high to be selected,
which prevents normal decoders like faad from being selected even in the
non digital case.