[tsdemux] rename PES pads to avoid name conflict with an existing one
When PMT is updated and the program switches,
the updated program is created/activated before deactivating the previous one.
If the both programs contain a PES of the same pid, this causes a name conflict
of the (new and old) pads for that PES, as pad name is derived from pid,
thus the new pad cannot be created or added to the element.