tsdemux,mpegtsbase: support PMT switch
to work inside decodebin,
swap the order of program add and remove in mpegtsbase,
because removing an old program before adding new one destructs
the old decoder group before a new decoder group is created,
which leads to EOS of the whole pipeline at PMT switches.

as a side effect, pad names must be renamed to avoid name conflicts
when the old and the new program contains same pids.